Separated networking, views and Models into their own files to make the code more readable (housekeeping after building a new feature)
